♻️ extract siteByDatacenter
