+++ scripts/lib/datadog-sites.js
@@ -1,0 +1,11 @@
+const siteByDatacenter = {
+  us1: 'datadoghq.com',
+  eu1: 'datadoghq.eu',
+  us3: 'us3.datadoghq.com',
+  us5: 'us5.datadoghq.com',
+  ap1: 'ap1.datadoghq.com',
+}
+
+module.exports = {
+  siteByDatacenter,
+}
